Modify(reseau) Optimize(reseau) New(plot)
Removed the plot method from lattice in reseau.py then reimplemented it as a standalone in plot.py.

Modified the electrostatic_energy methods to maximize the chances to form an exciton. (Still looking for the issue about these methods but it doesn't seem to happen anymore.)

Optimized events method. Now, the slower movements are pre-discarded. Witnessed a 2-3 times runtime reduction on the same example.