New(reseau) : cutoff radius
Implementation of cutoff radius for electrostatic energy.

Corrected an unanticipated behavior where a charge could be immobilized due to the new event creation algorithm. 
When an electron (respectively a hole) move into the neighbourhood of another one, there was a small chance that the event of the other charge was discarded because it was going to move to the same molecule. Now, it creates a new move event for the particule instead.