New(reseau) : decay event for each molecule type
Added methods for decay on the different molecules. Non-radiative decays for Host, fluorescence, phosphorescence, (R)ISC and FRET for TADF and fluorescence and non-radiative for Fluorophore.

Renamed Fluorescent class : Fluorophore. It is more consistent with the papers on (hyper)fluorescence.

Changed the operations method in Lattice so it stops when a certain amount of recombinations is reached or after a specified number of steps to prevent quasi-infinite loop.

*Still have to treat each events and debug but pretty much done.*