Change(reseau) : new lattice creation algorithm
Changed the lattice creation algorithm to stack layers of host at the bottom and the top of the lattice. This way, TADF and fluorophores are more concentrated.

Added the wave function coupling distance gamma to the calculation of move events.

Corrected a bug in plot.py due to the refactory of Fluorophore class.